Replace old registry impl by new one
